requirements.txt
Flask==2.1.2
halo==0.0.31
mysql_connector_repackaged==0.3.1
numpy==1.21.6
onnxruntime==1.11.1
onnxruntime_gpu==1.11.1
PyAudio==0.2.11
rx==3.2.0
SoundFile==0.10.3.post1
SpeechRecognition==3.8.1
torch==1.11.0
torchaudio==0.11.0
transformers==4.19.1
webrtcvad==2.0.10
